removed attributes from ConfigAPI
